JS-4716: remove double code for context menus
